+++ app/src/main/java/com/inky/fitnesscalendar/util/timeUtils.kt
@@ -2,10 +2,13 @@ package com.inky.fitnesscalendar.util
 
 import android.os.Build
 import com.inky.fitnesscalendar.data.EpochDay
+import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.ZoneId
+import java.time.temporal.WeekFields
 import java.util.Date
+import java.util.Locale
 
 fun LocalDateTime.toDate(zoneId: ZoneId = ZoneId.systemDefault()): Date =
     Date.from(atZone(zoneId).toInstant())
@@ -22,3 +25,8 @@ fun Date.toLocalDateTime(zoneId: ZoneId = ZoneId.systemDefault()): LocalDateTime
 
 fun Date.toEpochDay(zoneId: ZoneId = ZoneId.systemDefault()): EpochDay =
     EpochDay(toLocalDate(zoneId).toEpochDay())
+
+fun weekDays(locale: Locale): Sequence<DayOfWeek> {
+    val firstDayOfWeek = WeekFields.of(locale).firstDayOfWeek
+    return (0L..<7L).asSequence().map { firstDayOfWeek.plus(it) }
+}